Moving to OpenFace 2.0! Loads of changes.

+++ lib/local/CppInerop/OpenCVWrappers.h
@@ -181,7 +181,7 @@ namespace OpenCVWrappers {
 
 			int get()
 			{
-				return mat->step;
+				return (int) mat->step;
 			}
 		}
